[Refactor] Move rejected record sync daemon to DataWorkflows
