+++ be/module_boundary_manifest.json
@@ -616,6 +616,7 @@
         "runtime/",
         "compute_env/",
         "serde/",
+        "http/core/",
         "platform/",
         "fs/",
         "io/",
@@ -630,7 +631,6 @@
         "agent/",
         "service/",
         "connector/",
-        "http/core/",
         "http/service/",
         "script/",
         "tools/",
