+++ be/src/runtime/exec_env.cpp
@@ -76,7 +76,6 @@
 #include "runtime/lookup_stream_mgr.h"
 #include "runtime/mem_tracker.h"
 #include "runtime/pipeline_fragment_reporter.h"
-#include "runtime/rejected_record_sync_daemon.h"
 #include "runtime/routine_load/routine_load_task_executor.h"
 #include "runtime/runtime_filter_cache.h"
 #include "runtime/runtime_filter_worker.h"
@@ -307,23 +306,6 @@ Status ExecEnv::init(const std::vector<StorePath>& store_paths, ProcessMetricsRe
     if (!status.ok()) {
         LOG(ERROR) << "load path mgr init failed." << status.message();
         exit(-1);
-    }
-
-    // Phase 3 of the rejected records feature. The daemon thread is
-    // started unconditionally; its tick_loop() re-reads
-    // `config::enable_rejected_record_sync` every interval and treats a
-    // false value as a no-op. Starting it only when the flag is true at
-    // BE-boot time would break the mutable-config contract: operators
-    // (and tests) expect `update_config?enable_rejected_record_sync=true`
-    // to activate shipping without a BE restart. Leaving the thread
-    // parked costs one condvar wait per interval and no I/O when the
-    // flag is off.
-    _rejected_record_sync_daemon = new RejectedRecordSyncDaemon(this);
-    Status rr_status = _rejected_record_sync_daemon->init();
-    if (!rr_status.ok()) {
-        LOG(ERROR) << "RejectedRecordSyncDaemon init failed: " << rr_status.message();
-        // Non-fatal: the load path still works, we just don't ship
-        // rejected records to the system table this run.
     }
 
     StorageEnvOptions storage_env_options;
@@ -620,10 +602,6 @@ void ExecEnv::destroy() {
     SAFE_DELETE(_stream_load_executor);
     SAFE_DELETE(_connector_sink_spill_executor);
     SAFE_DELETE(_fragment_mgr);
-    if (_rejected_record_sync_daemon != nullptr) {
-        _rejected_record_sync_daemon->stop();
-    }
-    SAFE_DELETE(_rejected_record_sync_daemon);
     if (_compute_env != nullptr) {
         _compute_env->destroy_load_path();
     }
